refactor: rename workers to llm_apis and swaps to transformations
