+++ implementations/cryptohauntological_probe/llm_apis/claude_worker.py
@@ -1,14 +1,16 @@
 import os
-from openai import OpenAI
+import anthropic
 from typing import List, Tuple
 
-class OpenAIAPIWorker:
-    def __init__(self, api_key: str = None, model_name: str = "gpt-3.5-turbo"):
-        self.api_key = os.getenv("OPENAI_API_KEY")
+class ClaudeAPIWorker:
+    def __init__(self, api_key: str = None, model_name: str = "claude-3-opus-20240229", context_window_limit: int = 2048, timeout: int = 120):
+        self.api_key = os.getenv("CLAUDE_API_KEY")
         if not self.api_key:
-            raise ValueError("OpenAI API key must be provided via argument or OPENAI_API_KEY env var.")
-        self.client = OpenAI(api_key=self.api_key)
+            raise ValueError("Claude API key must be provided via argument or CLAUDE_API_KEY env var.")
+        self.client = anthropic.Anthropic(api_key=self.api_key)
         self.model_name = model_name
+        self.context_window_limit = context_window_limit
+        self.timeout = timeout
 
     def _build_history(self, history: List[Tuple[str, str]], user_query: str) -> list:
         messages = []
@@ -21,8 +23,9 @@ class OpenAIAPIWorker:
     def reply(self, prompt: str, memory: List[Tuple[str, str]] = None) -> str:
         memory = memory or []
         messages = self._build_history(memory, prompt)
-        chat_completion = self.client.chat.completions.create(
-            messages=messages,
+        message = self.client.messages.create(
             model=self.model_name,
+            max_tokens=1024,
+            messages=messages
         )
-        return chat_completion.choices[0].message.content
+        return message.content[0].text
