+++ implementations/cryptohauntological_probe/llm_apis/cohere_worker.py
@@ -3,12 +3,14 @@ import cohere
 from typing import List, Tuple
 
 class CohereAPIWorker:
-    def __init__(self, api_key: str = None, model_name: str = "command-r-plus"):
+    def __init__(self, api_key: str = None, model_name: str = "command-r-plus", context_window_limit: int = 2048, timeout: int = 120):
         self.api_key = os.getenv("COHERE_API_KEY")
         if not self.api_key:
             raise ValueError("Cohere API key must be provided via argument or COHERE_API_KEY env var.")
         self.client = cohere.Client(api_key=self.api_key)
         self.model_name = model_name
+        self.context_window_limit = context_window_limit
+        self.timeout = timeout
 
     def _build_history(self, history: List[Tuple[str, str]], user_query: str) -> list:
         messages = []
