+++ implementations/cryptohauntological_probe/llm_apis/gemini_worker.py
@@ -3,12 +3,14 @@ import google.generativeai as genai
 from typing import List, Tuple
 
 class GeminiAPIWorker:
-    def __init__(self, api_key: str = None, model_name: str = "models/gemini-pro"):
+    def __init__(self, api_key: str = None, model_name: str = "models/gemini-pro", context_window_limit: int = 2048, timeout: int = 120):
         self.api_key = os.getenv("GEMINI_API_KEY")
         if not self.api_key:
             raise ValueError("Gemini API key must be provided via argument or GEMINI_API_KEY env var.")
         genai.configure(api_key=self.api_key)
         self.model = genai.GenerativeModel(model_name)
+        self.context_window_limit = context_window_limit
+        self.timeout = timeout
 
     def _build_history(self, history: List[Tuple[str, str]], user_query: str) -> list:
         messages = []
