+++ implementations/cryptohauntological_probe/llm_apis/mistral_worker.py
@@ -3,12 +3,14 @@ from mistralai.client import MistralClient
 from typing import List, Tuple
 
 class MistralAPIWorker:
-    def __init__(self, api_key: str = None, model_name: str = "mistral-large-latest"):
+    def __init__(self, api_key: str = None, model_name: str = "mistral-large-latest", context_window_limit: int = 2048, timeout: int = 120):
         self.api_key = os.getenv("MISTRAL_API_KEY")
         if not self.api_key:
             raise ValueError("Mistral API key must be provided via argument or MISTRAL_API_KEY env var.")
         self.client = MistralClient(api_key=self.api_key)
         self.model_name = model_name
+        self.context_window_limit = context_window_limit
+        self.timeout = timeout
 
     def _build_history(self, history: List[Tuple[str, str]], user_query: str) -> list:
         messages = []
