+++ implementations/cryptohauntological_probe/llm_apis/ollama_worker.py
@@ -1,14 +1,11 @@
-import os
-import anthropic
+import ollama
 from typing import List, Tuple
 
-class ClaudeAPIWorker:
-    def __init__(self, api_key: str = None, model_name: str = "claude-3-opus-20240229"):
-        self.api_key = os.getenv("CLAUDE_API_KEY")
-        if not self.api_key:
-            raise ValueError("Claude API key must be provided via argument or CLAUDE_API_KEY env var.")
-        self.client = anthropic.Anthropic(api_key=self.api_key)
+class OllamaAPIWorker:
+    def __init__(self, model_name: str = "llama3", context_window_limit: int = 2048, timeout: int = 120):
         self.model_name = model_name
+        self.context_window_limit = context_window_limit
+        self.timeout = timeout
 
     def _build_history(self, history: List[Tuple[str, str]], user_query: str) -> list:
         messages = []
@@ -21,9 +18,5 @@ class ClaudeAPIWorker:
     def reply(self, prompt: str, memory: List[Tuple[str, str]] = None) -> str:
         memory = memory or []
         messages = self._build_history(memory, prompt)
-        message = self.client.messages.create(
-            model=self.model_name,
-            max_tokens=1024,
-            messages=messages
-        )
-        return message.content[0].text
+        response = ollama.chat(model=self.model_name, messages=messages)
+        return response['message']['content']
