+++ implementations/cryptohauntological_probe/llm_apis/openai_worker.py
@@ -1,14 +1,16 @@
 import os
-from groq import Groq
+from openai import OpenAI
 from typing import List, Tuple
 
-class GroqAPIWorker:
-    def __init__(self, api_key: str = None, model_name: str = "llama3-8b-8192"):
-        self.api_key = os.getenv("GROQ_API_KEY")
+class OpenAIAPIWorker:
+    def __init__(self, api_key: str = None, model_name: str = "gpt-3.5-turbo", context_window_limit: int = 2048, timeout: int = 120):
+        self.api_key = os.getenv("OPENAI_API_KEY")
         if not self.api_key:
-            raise ValueError("Groq API key must be provided via argument or GROQ_API_KEY env var.")
-        self.client = Groq(api_key=self.api_key)
+            raise ValueError("OpenAI API key must be provided via argument or OPENAI_API_KEY env var.")
+        self.client = OpenAI(api_key=self.api_key)
         self.model_name = model_name
+        self.context_window_limit = context_window_limit
+        self.timeout = timeout
 
     def _build_history(self, history: List[Tuple[str, str]], user_query: str) -> list:
         messages = []
